requirements.txt
# Core Framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
gunicorn>=21.2.0
pydantic>=2.6.3
email_validator==2.2.0
pytz==2025.1

# Database
SQLAlchemy>=2.0.27
psycopg2-binary>=2.9.9
redis>=5.0.1

# Authentication & Security
PyJWT==2.10.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
python-multipart>=0.0.9

# API & Networking
requests>=2.31.0
websockets>=12.0
python-decouple>=3.8

# AI Integration
openai>=1.12.0
